Python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
antlr4-python3-runtime==4.9.3
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
docopt==0.6.2
einops==0.8.1
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.1
huggingface-hub==0.29.3
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.8
langdetect==1.0.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.44.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
omegaconf==2.3.0
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
optuna==4.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
praat-parselmouth==0.4.5
primePy==1.3
propcache==0.3.1
protobuf==6.30.2
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.40
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.6.0
torchmetrics==1.7.0
tqdm==4.67.1
transformers==4.50.3
typer==0.15.2
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Wave==0.0.2
whisper==1.1.10
yarl==1.18.3
